requirements.txt
torch
torchvision
transformers
huggingface_hub
numpy
matplotlib
opencv-python
Pillow
ftfy
regex
seaborn
scipy